requirements.txt
Flask==2.0.1
torch==1.9.0
torchvision==0.10.0
mediapipe==0.8.5
opencv-python==4.5.3.20210927
Pillow==8.4.0
requests==2.25.1
